+++ .gitignore
@@ -128,4 +128,3 @@ Thumbs.db
 # Project specific
 # Add any project specific files to ignore here
 job_search_assistant.txt
-plan.md
