+++ src/agent/workflows/job_evaluation/__init__.py
@@ -5,6 +5,12 @@ This module contains the workflow for evaluating job postings against
 user criteria, including extraction and evaluation steps.
 """
 
+from src.agent.workflows.job_evaluation.main import (
+    JobEvaluationWorkflow,
+    create_job_evaluation_workflow,
+    evaluate_job_posting,
+    generate_recommendation_from_results,
+)
 from src.agent.workflows.job_evaluation.states import (
     JobEvaluationWorkflowState,
     add_message,
@@ -12,13 +18,10 @@ from src.agent.workflows.job_evaluation.states import (
     get_extracted_info_as_dict,
     is_extraction_successful,
 )
-from src.agent.workflows.job_evaluation.workflow import (
-    JobEvaluationWorkflow,
-    create_job_evaluation_workflow,
-    generate_recommendation_from_results,
-)
 
 __all__ = [
+    # Public API
+    "evaluate_job_posting",
     # State management
     "JobEvaluationWorkflowState",
     "create_initial_state",
