+++ src/agent/workflows/job_evaluation/main.py
@@ -1,8 +1,11 @@
 """
 Defines the job evaluation workflow using a StateGraph.
+
+This module provides the complete job evaluation workflow implementation,
+including the public API function that serves as the main entry point.
 """
 
-from typing import Any, Optional
+from typing import Any, Dict, Optional
 
 from langgraph.graph import END, StateGraph
 
@@ -12,6 +15,7 @@ from src.agent.agents.extraction.job_extraction_agent import (
     validation_node,
 )
 from src.agent.workflows.job_evaluation.states import JobEvaluationWorkflowState
+from src.llm.langfuse_handler import get_langfuse_handler
 from src.utils.logging import get_logger
 
 logger = get_logger(__name__)
@@ -77,14 +81,11 @@ class JobEvaluationWorkflow:
         return final_state
 
 
-# Backward compatibility functions for existing code
 def generate_recommendation_from_results(
     evaluation_results: dict,
 ) -> tuple[str, str]:
     """
     Generate final recommendation and reasoning from evaluation results.
-
-    This function maintains backward compatibility with existing code.
 
     Args:
         evaluation_results: Dictionary with evaluation results for each criteria
@@ -131,3 +132,79 @@ def create_job_evaluation_workflow() -> JobEvaluationWorkflow:
         JobEvaluationWorkflow: New workflow instance
     """
     return JobEvaluationWorkflow()
+
+
+def evaluate_job_posting(job_posting_text: str) -> Dict[str, Any]:
+    """
+    Evaluates a job posting using the graph-based agent workflow.
+
+    This is the main public API for the job evaluation system. It provides
+    a clean interface to the underlying LangGraph workflow system.
+
+    Args:
+        job_posting_text: The raw job posting text to evaluate.
+
+    Returns:
+        A dictionary with evaluation results in the format expected by the UI:
+        - recommendation: "APPLY", "DO_NOT_APPLY", or "ERROR"
+        - reasoning: Human-readable explanation of the recommendation
+        - extracted_info: Dictionary with structured job information
+        - evaluation_result: Dictionary with detailed evaluation criteria results
+    """
+    logger.info("Starting job evaluation with graph-based workflow.")
+
+    if not job_posting_text or not job_posting_text.strip():
+        logger.warning("Job posting text is empty.")
+        return {
+            "recommendation": "ERROR",
+            "reasoning": "Job posting text was empty.",
+            "extracted_info": {},
+            "evaluation_result": {},
+        }
+
+    langfuse_handler = get_langfuse_handler()
+
+    try:
+        workflow = JobEvaluationWorkflow()
+        final_state = workflow.run(job_posting_text, langfuse_handler)
+
+        result = final_state.get("evaluation_result")
+        if not result:
+            logger.error("Workflow completed but no evaluation result was generated.")
+            return {
+                "recommendation": "ERROR",
+                "reasoning": "Workflow failed to generate an evaluation result.",
+                "extracted_info": final_state.get("extracted_info", {}),
+                "evaluation_result": {},
+            }
+
+        recommendation, reasoning = generate_recommendation_from_results(result)
+
+        # Convert extracted_info to dictionary format for UI compatibility
+        extracted_info_dict = {}
+        extracted_info = final_state.get("extracted_info")
+        if extracted_info:
+            if hasattr(extracted_info, "model_dump"):
+                extracted_info_dict = extracted_info.model_dump()
+            elif isinstance(extracted_info, dict):
+                extracted_info_dict = extracted_info
+            else:
+                extracted_info_dict = dict(extracted_info) if extracted_info else {}
+
+        logger.info(f"Job evaluation completed with recommendation: {recommendation}")
+
+        return {
+            "recommendation": recommendation,
+            "reasoning": reasoning,
+            "extracted_info": extracted_info_dict,
+            "evaluation_result": result,
+        }
+
+    except Exception as e:
+        logger.error(f"Job evaluation workflow failed: {e}", exc_info=True)
+        return {
+            "recommendation": "ERROR",
+            "reasoning": f"An unexpected error occurred during the workflow: {e}",
+            "extracted_info": {},
+            "evaluation_result": {},
+        }
